requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.3
django-bootstrap==0.2.4
django-bootstrap4==1.0.1
django-leaflet==0.26.0
django-recaptcha2==1.4.1
djangorestframework==3.10.3
idna==2.9
Pillow==6.2.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
